backend/requirements.txt
uvicorn[standard]
fastapi
pymongo
pydantic
websockets
python-jose
passlib[bcrypt]
pytest_asyncio
python-decouple
motor
bcrypt
httpx
mongomock_motor
pytest
python-multipart
types-python-jose
google-genai